receipts and category apis implementation

+++ e-commerce/project/store/views.py
@@ -33,6 +33,7 @@ def register(request):
 
 @api_view(['POST'])
 def login(request):
+    print("hello")
     serializer = UserSerializer(data=request.data)
     print(serializer.initial_data)
     print(serializer.is_valid())
@@ -41,7 +42,7 @@ def login(request):
     data["token"] = token
     return Response(data)
 
-
+# after editing you have to login again.
 @api_view(['POST'])
 @permission_classes((IsAuthenticated,))
 def edit_info(request, pk):
@@ -50,56 +51,62 @@ def edit_info(request, pk):
     usr.delete()
     if serializer.is_valid():
         serializer.save()
-        return Response(serializer.data)
+        return Response("edited succesfully")
+    return Response("you don't have premission")
 
 
 
 @api_view(['POST'])
 @permission_classes((IsAuthenticated,))
 def add_category(request):
-    serializer = CategorySerializer(data=request.data)
-    print(serializer.is_valid())
-    if serializer.is_valid():
-        serializer.save()
-    return Response(serializer.data)
+    user = request.user
+    if user.is_admin:
+        serializer = CategorySerializer(data=request.data)
+        print(serializer.is_valid())
+        if serializer.is_valid():
+            serializer.save()
+        return Response("category added")
+    return Response("you don't have premission")
 
 
 
 @api_view(['POST'])
 @permission_classes((IsAuthenticated,))
 def update_category(request, pk):
-    cat = category.objects.get(category_name=pk)
-    serializer = CategorySerializer(instance=cat, data=request.data)
-    if serializer.is_valid():
-        cat.delete()
-        serializer.save()
-    return Response(serializer.data)
+    if request.user.is_admin:
+        cat = category.objects.get(category_name=pk)
+        serializer = CategorySerializer(instance=cat, data=request.data)
+        if serializer.is_valid():
+            cat.delete()
+            serializer.save()
+        return Response("category updated")
+    return Response("you don't have premission")
 
 
 @api_view(['PUT'])
 @permission_classes((IsAuthenticated,))
 def delete_category(request):
-    serializer = CategorySerializer(data=request.data)
-    cat = category.objects.get(category_name=serializer.initial_data["category_name"])
-    cat.delete()
-    return Response(serializer.initial_data)
+    if request.user.is_admin:
+        serializer = CategorySerializer(data=request.data)
+        cat = category.objects.get(category_name=serializer.initial_data["category_name"])
+        cat.delete()
+        return Response("category update")
+    return Response("you don't have premission")
 
-
-# def print_cat(request):
-#     for i in category.objects.all():
-#         print(i.category_name)
-#     return HttpResponse('{"hello world ":"hlo"}')
 
 
 
 @api_view(['GET'])
 @permission_classes((IsAuthenticated,))
-def receipts(resquest):
-    # r = receipt(tracing_code="fs3hfwjkl4j", stuff_name = "kala", user_name = "amjd@jfdsk.jin" , first_name = "amin", last_name = "rahimi", address = "theran", creation_date="2000-02-02", state="انجام شده")
-    # r.save()
-    rec = receipt.objects.all()
-    serializer = ReceiptSerializer(rec, many=True)
-    return Response(serializer.data)
+def receipts(request):
+    if request.user.is_admin:
+        rec = receipt.objects.all()
+        serializer = ReceiptSerializer(rec, many=True)
+        return Response(serializer.data)
+    else:
+        rec = receipt.objects.filter(user_name=request.user.user_name)
+        serializer = ReceiptSerializer(rec, many=True)
+        return Response(serializer.data)    
 
 
 @api_view(['GET'])
